wip strict

+++ src/asset_serialization.h
@@ -30,6 +30,6 @@ namespace AssetSerialization {
 	void InitializeMetadataJson(nlohmann::json& json, u64 id);
 	SerializationResult CreateAssetMetadataFile(const std::filesystem::path& path, u64 guid, nlohmann::json& outMetadata);
 
-	SerializationResult LoadAssetFromFile(const std::filesystem::path& path, AssetType type, const nlohmann::json& metadata, u32& size, void* pOutData);
+	SerializationResult LoadAssetFromFile(const std::filesystem::path& path, AssetType type, const nlohmann::json& metadata, size_t& size, void* pOutData);
 	SerializationResult SaveAssetToFile(const std::filesystem::path& path, const char* name, AssetType type, nlohmann::json& metadata, const void* pData);
 }
